server/hubco/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
dj-database-url==0.5.0
Django==3.2.5
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.0
openpyxl==3.0.7
pandas==1.3.0
protobuf==3.17.3
psycopg2==2.9.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0